# survey-parent/api-survey/src/main/resources/static/json/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "survey",
    "version": "v1.0",
    "title": "API文档",
    "termsOfService": "lego.cn",
    "contact": {
      "name": "lego",
      "url": "lego.cn",
      "email": "survey@legocloud.cn"
    }
  },
  "host": "192.168.101.197:50010",
  "basePath": "/",
...
  "paths": {
...
    "/survey/api-project/section/create": {
      "post": {
        "tags": [
          "section-controller"
        ],
        "summary": "添加标段信息",
        "description": "添加标段信息",
        "operationId": "addUsingPOST_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "in": "body",
            "name": "section",
            "description": "section",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Section"
            }
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-project/section/delete": {
      "delete": {
        "tags": [
          "section-controller"
        ],
        "summary": "删除标段信息",
        "description": "删除标段信息",
        "operationId": "deleteUsingDELETE_4",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "id",
            "in": "query",
            "description": "标段id",
            "required": true,
            "type": "string"
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-project/section/list/{pageIndex}": {
      "get": {
        "tags": [
          "section-controller"
        ],
        "summary": "获取标段列表",
        "description": "获取标段列表",
        "operationId": "listUsingGET_5",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "pageIndex",
            "in": "path",
            "description": "当前页",
            "required": true,
            "type": "integer",
            "format": "int32",
            "x-example": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页大小",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32",
            "x-example": 10
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO«RespDataVO«SectionVo»»"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-project/section/modify": {
      "put": {
        "tags": [
          "section-controller"
        ],
        "summary": "修改标段信息",
        "description": "修改标段信息",
        "operationId": "modifyUsingPUT_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "in": "body",
            "name": "section",
            "description": "section",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Section"
            }
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "ApiError": {
      "type": "object",
      "title": "ApiError"
    },
...
    "ColleaguesVo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "测量员id"
        },
        "name": {
          "type": "string",
          "description": "测量员昵称"
        },
        "phone": {
          "type": "string",
          "description": "测量员手机号"
        },
        "role": {
          "type": "string",
          "description": "测量员角色"
        }
      },
      "title": "ColleaguesVo"
    },
...
    "GroupVo": {
      "type": "object",
      "properties": {
        "desc": {
          "type": "string",
          "description": "单位描述"
        },
        "id": {
          "type": "string",
          "description": "单位ID"
        },
        "name": {
          "type": "string",
          "description": "单位名称"
        },
        "upperGroup": {
          "description": "上级单位",
          "$ref": "#/definitions/UpperGroup"
        }
      },
      "title": "GroupVo"
    },
...
    "Master": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "标段负责人id"
        },
        "name": {
          "type": "string",
          "description": "标段负责人名称"
        }
      },
      "title": "Master"
    },
...
    "OwnWorkspace": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "description": "工区编号"
        },
        "desc": {
          "type": "string",
          "description": "工区描述"
        },
        "id": {
          "type": "string",
          "description": "所属工区id"
        },
        "name": {
          "type": "string",
          "description": "所属工区名称"
        },
        "surveyer": {
          "type": "array",
          "description": "工区所属测量员",
          "items": {
            "$ref": "#/definitions/Surveyer"
          }
        },
        "type": {
          "type": "string",
          "description": "工区类型"
        },
        "valid": {
          "type": "integer",
          "format": "int32",
          "description": "是否有效 0-是;1-否"
        }
      },
      "title": "OwnWorkspace"
    },
    "OwnerProject": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "description": "工程编号"
        },
        "id": {
          "type": "string",
          "description": "所属工程id"
        },
        "name": {
          "type": "string",
          "description": "所属工程名称"
        }
      },
      "title": "OwnerProject"
    },
...
    "RespDataVO«SectionVo»": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SectionVo"
          }
        }
      },
      "title": "RespDataVO«SectionVo»"
    },
...
    "RespVO": {
      "type": "object",
      "properties": {
        "info": {
          "type": "object"
        },
        "msg": {
          "type": "string"
        },
        "ret": {
          "type": "string"
        },
        "retCode": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "RespVO"
    },
...
    "RespVO«RespDataVO«SectionVo»»": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/RespDataVO«SectionVo»"
        },
        "msg": {
          "type": "string"
        },
        "ret": {
          "type": "string"
        },
        "retCode": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "RespVO«RespDataVO«SectionVo»»"
    },
...
    "Section": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "description": "标段编号"
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "新增时间"
        },
        "desc": {
          "type": "string",
          "description": "标段描述"
        },
        "id": {
          "type": "string",
          "description": "标段id"
        },
        "master": {
          "type": "array",
          "description": "标段负责人",
          "items": {
            "$ref": "#/definitions/Master"
          }
        },
        "name": {
          "type": "string",
          "description": "标段名称"
        },
        "ownerGroup": {
          "description": "标段所属单位id",
          "$ref": "#/definitions/UpperGroup"
        },
        "ownerProject": {
          "description": "标段所属项目id",
          "$ref": "#/definitions/OwnerProject"
        },
        "property": {
          "type": "array",
          "description": "标段起始里程，标段结束里程",
          "items": {
            "type": "number",
            "format": "double"
          }
        },
        "service": {
          "type": "array",
          "description": "场景",
          "items": {
            "type": "string"
          }
        },
        "surveyer": {
          "type": "array",
          "description": "标段测量员id",
          "items": {
            "$ref": "#/definitions/Surveyer"
          }
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        },
        "valid": {
          "type": "integer",
          "format": "int32",
          "description": "是否有效 0-是;1-否"
        },
        "workSpace": {
          "type": "array",
          "description": "标段下的工区",
          "items": {
            "$ref": "#/definitions/OwnWorkspace"
          }
        }
      },
      "title": "Section"
    },
    "SectionVo": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "description": "标段编号"
        },
        "colleagues": {
          "type": "array",
          "description": "标段包含的测量员",
          "items": {
            "$ref": "#/definitions/ColleaguesVo"
          }
        },
        "endMileage": {
          "type": "number",
          "format": "double",
          "description": "标段结束里程"
        },
        "group": {
          "description": "标段所属单位",
          "$ref": "#/definitions/GroupVo"
        },
        "id": {
          "type": "string",
          "description": "标段id"
        },
        "name": {
          "type": "string",
          "description": "标段名称"
        },
        "services": {
          "type": "array",
          "description": "场景",
          "items": {
            "type": "string"
          }
        },
        "startingMileage": {
          "type": "number",
          "format": "double",
          "description": "标段起始里程"
        },
        "workspace": {
          "type": "array",
          "description": "标段包含的工区",
          "items": {
            "$ref": "#/definitions/WorkspaceVo"
          }
        }
      },
      "title": "SectionVo"
    },
...
    "Surveyer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "测量员id"
        },
        "name": {
          "type": "string",
          "description": "测量员名称"
        }
      },
      "title": "Surveyer"
    },
...
    "UpperGroup": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "上级单位id"
        },
        "name": {
          "type": "string",
          "description": "上级单位名称"
        }
      },
      "title": "UpperGroup"
    },
...
    "WorkspaceVo": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "description": "工区编码"
        },
        "id": {
          "type": "string",
          "description": "工区id"
        },
        "name": {
          "type": "string",
          "description": "工区名称"
        },
        "permission": {
          "type": "string",
          "description": "工区权限"
        },
        "type": {
          "type": "string",
          "description": "工区类型"
        }
      },
      "title": "WorkspaceVo"
    }
  }
}